requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
openai
httpx